It was improved the delete functionality to delete the info file.

+++ src/br/edu/ifsp/util/Delete.java
@@ -8,11 +8,13 @@ public class Delete {
 
 	public void deleteFile(File file) throws IOException {
 		File files[] = file.listFiles();
-		for (File f : files) {
-			if (f.isDirectory()) {
-				this.deleteFile(f);
-			} else if (f.isFile()) {
-				Files.delete(f.toPath());
+		if (files != null) {
+			for (File f : files) {
+				if (f.isDirectory()) {
+					this.deleteFile(f);
+				} else if (f.isFile()) {
+					Files.delete(f.toPath());
+				}
 			}
 		}
 		Files.delete(file.toPath());
@@ -23,21 +25,26 @@ public class Delete {
 		File color = new File(file.getAbsolutePath() + File.separator + "Color");
 		File coord = new File(file.getAbsolutePath() + File.separator + "Coordinates");
 		File seg = new File(file.getAbsolutePath() + File.separator + "Segmentation");
-		
-		if(depth.exists()){
+		File info = new File(file.getAbsolutePath() + File.separator + "info.json");
+
+		if (depth.exists()) {
 			deleteFile(depth);
 		}
-		
-		if(color.exists()){
+
+		if (color.exists()) {
 			deleteFile(color);
 		}
-		
-		if(coord.exists()){
+
+		if (coord.exists()) {
 			deleteFile(coord);
 		}
-		
-		if(seg.exists()){
+
+		if (seg.exists()) {
 			deleteFile(seg);
+		}
+
+		if (info.exists()) {
+			deleteFile(info);
 		}
 	}
 }
